cloud-functions-svc/requirements.txt
functions-framework==3.*
kafka-python
google-cloud-bigquery
google-cloud-secret-manager
requests